Split the command menus
