+++ examples/r4-command-menu/index.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html>
+	<head>
+		<meta charset="utf-8" />
+		<meta name="viewport" content="width=device-width" />
+		<link rel="icon" href="data:;base64,iVBORw0KGgo=" />
+		<title>r4-command-menu</title>
+		<meta name="description" content="r4-avatar" />
+		<link rel="stylesheet" href="/styles/index.css" />
+		<script type="module" src="/src/index.js"></script>
+	</head>
+
+	<body>
+		<p>r4-command-menu</p>
+
+		<p>There are two ways to tell the component what to render:</p>
+		<ul>
+			<li>Set the `image` attribute to a Cloudinary image id</li>
+			<li>Set the `slug` attribute to a channel slug. This will cause a network request.</li>
+		</ul>
+		<r4-command-menu></r4-command-menu>
+	</body>
+</html>
